Reorganize project into cmd internal and pkg
