+++ internal/chat/errors.go
