+++ internal/chat/user/sender.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/gorilla/websocket"
 
-	"technochat/chat/message"
+	"technochat/internal/chat/message"
 )
 
 const (
