+++ internal/http/messageview.go
@@ -5,7 +5,7 @@ import (
 	"html"
 	"log"
 	"net/http"
-	"technochat/entity"
+	"technochat/pkg/entity"
 )
 
 type messageViewRequest struct {
