+++ pkg/entity/errors.go
