+++ pkg/entity/imagesarray.go
