+++ pkg/entity/message.go
